move sortversion.py out of lib
